# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"rag-knowledge","version":"0.1.0"},"paths":{"/api/health":{"get":{"tags":["health"],"summary":"Health","operationId":"healthApp","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthStatus"}}}}}}},"/api/health/db":{"get":{"tags":["health"],"summary":"Health Db","operationId":"healthDb","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthStatus"}}}}}}},"/api/health/cos":{"get":{"tags":["health"],"summary":"Health Cos","operationId":"healthCos","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthStatus"}}}}}}},"/api/documents":{"post":{"tags":["documents"],"summary":"Upload Document","description":"上传文档：写入 COS、落库后立即返回，解析与向量化通过 BackgroundTasks 异步进行。","operationId":"uploadDocument","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_uploadDocument"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["documents"],"summary":"List Documents","operationId":"listDocuments","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Page Size"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"enum":["uploading","parsing","indexing","ready","failed"],"type":"string"},{"type":"null"}],"description":"按文档状态筛选","title":"Status"},"description":"按文档状态筛选"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/{document_id}":{"get":{"tags":["documents"],"summary":"Get Document","operationId":"getDocument","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Document Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["documents"],"summary":"Delete Document","operationId":"deleteDocument","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Document Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/{document_id}/retry":{"post":{"tags":["documents"],"summary":"Retry Document","operationId":"retryDocument","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Document Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/{document_id}/file":{"get":{"tags":["documents"],"summary":"Download Document","description":"返回文档原始字节。\n\n- PDF / HTML / Markdown: 可在浏览器内联预览\n- DOCX: 浏览器无法渲染, 强制 attachment","operationId":"downloadDocument","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Document Id"}},{"name":"download","in":"query","required":false,"schema":{"type":"integer","maximum":1,"minimum":0,"description":"1=强制下载, 0=尝试内联预览","default":0,"title":"Download"},"description":"1=强制下载, 0=尝试内联预览"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/{document_id}/chunks":{"get":{"tags":["documents"],"summary":"List Document Chunks","operationId":"listDocumentChunks","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Document Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentChunkListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/{document_id}/chunks/{chunk_id}":{"get":{"tags":["documents"],"summary":"Get Document Chunk","operationId":"getDocumentChunk","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Document Id"}},{"name":"chunk_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Chunk Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentChunkDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_uploadDocument":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File","description":"待上传文档（PDF / DOCX / Markdown / HTML）"}},"type":"object","required":["file"],"title":"Body_uploadDocument"},"DocumentChunkDetail":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"document_id":{"type":"string","format":"uuid","title":"Document Id"},"chunk_index":{"type":"integer","title":"Chunk Index"},"page_no":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page No"},"section_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Section Path"},"content":{"type":"string","title":"Content"},"char_count":{"type":"integer","title":"Char Count"},"chunk_hash":{"type":"string","title":"Chunk Hash"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","document_id","chunk_index","content","char_count","chunk_hash","created_at"],"title":"DocumentChunkDetail","description":"chunk 详情：返回完整 content。"},"DocumentChunkListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/DocumentChunkRead"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","minimum":1.0,"title":"Page"},"page_size":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"Page Size"},"stats":{"anyOf":[{"$ref":"#/components/schemas/DocumentChunkStats"},{"type":"null"}]}},"type":"object","required":["items","total","page","page_size"],"title":"DocumentChunkListResponse"},"DocumentChunkRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"chunk_index":{"type":"integer","title":"Chunk Index"},"page_no":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page No"},"section_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Section Path"},"content_excerpt":{"type":"string","title":"Content Excerpt"},"char_count":{"type":"integer","title":"Char Count"},"chunk_hash":{"type":"string","title":"Chunk Hash"}},"type":"object","required":["id","chunk_index","content_excerpt","char_count","chunk_hash"],"title":"DocumentChunkRead","description":"chunk 列表项。content_excerpt 已在 API 层截断到固定长度。"},"DocumentChunkStats":{"properties":{"total":{"type":"integer","title":"Total"},"avg_length":{"type":"integer","title":"Avg Length"},"min_length":{"type":"integer","title":"Min Length"},"max_length":{"type":"integer","title":"Max Length"}},"type":"object","required":["total","avg_length","min_length","max_length"],"title":"DocumentChunkStats","description":"切分统计：直观看到 chunk_size / overlap 配置的实际效果。"},"DocumentListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/DocumentRead"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","minimum":1.0,"title":"Page"},"page_size":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"Page Size"}},"type":"object","required":["items","total","page","page_size"],"title":"DocumentListResponse"},"DocumentRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"file_hash":{"type":"string","title":"File Hash"},"mime_type":{"type":"string","title":"Mime Type"},"size":{"type":"integer","title":"Size"},"status":{"type":"string","enum":["uploading","parsing","indexing","ready","failed"],"title":"Status"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","name","file_hash","mime_type","size","status","created_at","updated_at"],"title":"DocumentRead"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthStatus":{"properties":{"status":{"type":"string","enum":["ok","error","not_configured"],"title":"Status"},"detail":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Detail"}},"type":"object","required":["status"],"title":"HealthStatus"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}